Remove unneeded if, clean up types
